A serial interface that using RS-232 protocol is included for communication between rfid reader and microcontroller
A Serial RS-232 output which makes it easy to interface it to any
microcontroller or computer. The reader senses any RFID tag within a
range of 50mm to 150mm and outputs the tag data through both serial
RS-232 connection and through TTL pins. The demodulated serial data is
compared to existing tag number and if there any number matching with
the received number then identifies presence of that object/vehicle, and
deducts the ticket amount from their account.